requirements.txt
Flask==2.0.2
python-dotenv==0.19.2
requests==2.25.1
gunicorn==20.1.0
Unidecode==1.3.2

urllib3==1.26.9
facebook-sdk==3.1.0